.key-feature__image
keys notes to remember is that to get a perfect circle, your width and height of the properties must be the same (to make a square) and then use border radius: 50% to make a circle.
to align the .key-feature__image to the center use the property "margin" and set the value to 0 auto. this will align the item center horizontally. this will do nothing if you want to center it vertically.