package page
new  pages title "package" i copied everything  from the root .html to the package sub-folder with the exception of the main tag. I also, changed the name of the links by adding ../ since  im working on the subfolder, i have to use ../ to access the other folders when im in this subfolder.