shared.js
for () loop are used to repeat code which i am using to pass 3 arguments.

the first is the iteration variable (i which is set top 0), a temporary variable which i use to keep track how often the loop is executed .

then we add the condition for this loop to determine how long this loop would keep running  (i < selectPlanButtons.length) this tells us that it should keep running as long as "i" is smaller than the select plan buttons length. (length is a special property which tell us how many elements we have)
var i = 0; i < selectPlanButtons.length;
this will keep running until there is 2. for 2 it will still be executed  but once "i: reaches 3, it will not execute the code again because three is not smaller than three.

the third argument is seperated by a semicolon which is essentially just the code which shoukd run whenever we finished one loop iteration, it shoud increment "i" by one (i++) this will add 1 to "i" each run through that loop.