closing the modal.
target the only button in the class modal  or target the name that targets the button which is .modal__action--negative.

add a new variable which i would call modalNoButton then use document.querrySelector and add the .modal__action--negative selector.

add event listener to the variable backdrop and modalNoButton
afterwards we pass a function whenever this event occurs.

backdrop.addEventListener('click', function () {
   modal.style.display = 'none';
    backdrop.style.display = 'none';
});

copy and add the same code to modalNoButton and since its the same code, we can outsource the function and name it "closeModal"

function closeModal() {
    modal.style.display = 'none';
    backdrop.style.display = 'none';
}

and with that, closeModal is now a named function which we can use as a second argument instead of defining the function in the backdrop and modalNoButton