restyling form element
 we made some work on restyling the inputs and ensured that they use our general font anf above all maniputale their style.

side note: by deafuault the input element has an "outline" property applied to it but we get rid of that by adding the psuedo element :focus to target the element when its focused and setting ouitline to "none". the outline always comes after the border so you can have both, it is added outside of the border.