styling check-box
I want to ensure that the style that targets all inputs, which includes the checkbox,
do not target the check box. to ensure that its not targeted, we'll target the ".signup-form input,
.signup-form select" selector and add the ":not()" pseudo selector. this pseudo selector allows us to say that it should
target all input which do not fulfill the requirements we add here inside the parenthesis. ill add another attribute selector
with square brackets type="checkbox".

by deafult the checkbox is a special type of input in which it has a predetermained styling and to overwright that, you will need
put it in your selector -webkit-appearance: none; -moz-appearance: none; appearance: none; we are forcing the webbrowserr to accept
our style. this, will not make the ability to check the checkbox unless we add the psuedo element :checked