user feedback
the manual approach is to simply add a class"invalid" on any condition defined by you, wither on a server or through javascript and you just simply give that class the invalid styling you want to have such as:
.signup-form input.invalid,
.signup-form select.invalid {
    border-color: rgb(250, 172, 172);
    background: rgb(250, 172, 172);
}

there are multiple ways to go about it but a more elegant way to achieve usser feedback is to use the "invalid pseudo selector. If  you want to validate the other input too, you can add "required" attribute which tells the browser that "this is invalid if its empty" theres also a valid pseudo element which does the opposite, anything that is valid will have said style.